requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
Office365-REST-Python-Client>=2.4.0
python-dateutil>=2.8.0